+++ spring-grpc-spring-boot-autoconfigure/src/test/java/org/springframework/grpc/autoconfigure/client/GrpcClientPropertiesTests.java
@@ -17,19 +17,23 @@
 package org.springframework.grpc.autoconfigure.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.entry;
 
 import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
 
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.context.properties.bind.Binder;
 import org.springframework.boot.context.properties.source.MapConfigurationPropertySource;
-import org.springframework.grpc.client.NamedChannel;
+import org.springframework.grpc.autoconfigure.client.GrpcClientProperties.ChannelConfig;
 import org.springframework.grpc.client.NegotiationType;
+import org.springframework.mock.env.MockEnvironment;
+import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.util.unit.DataSize;
 
 /**
@@ -45,6 +49,13 @@ class GrpcClientPropertiesTests {
 			.get();
 	}
 
+	private GrpcClientProperties newProperties(ChannelConfig defaultChannel, Map<String, ChannelConfig> channels) {
+		var properties = new GrpcClientProperties();
+		ReflectionTestUtils.setField(properties, "defaultChannel", defaultChannel);
+		ReflectionTestUtils.setField(properties, "channels", channels);
+		return properties;
+	}
+
 	@Nested
 	class BindPropertiesAPI {
 
@@ -55,11 +66,11 @@ class GrpcClientPropertiesTests {
 
 		@Test
 		void specificChannelWithDefaultValues() {
-			this.withDefaultValues("channels.c1", (p) -> p.getChannels().get("c1"));
+			this.withDefaultValues("channels.c1", (p) -> p.getChannel("c1"));
 		}
 
 		private void withDefaultValues(String channelName,
-				Function<GrpcClientProperties, NamedChannel> channelFromProperties) {
+				Function<GrpcClientProperties, ChannelConfig> channelFromProperties) {
 			Map<String, String> map = new HashMap<>();
 			// we have to at least bind one property or bind() fails
 			map.put("spring.grpc.client.%s.enable-keep-alive".formatted(channelName), "false");
@@ -91,11 +102,11 @@ class GrpcClientPropertiesTests {
 
 		@Test
 		void specificChannelWithSpecifiedValues() {
-			this.withSpecifiedValues("channels.c1", (p) -> p.getChannels().get("c1"));
+			this.withSpecifiedValues("channels.c1", (p) -> p.getChannel("c1"));
 		}
 
 		private void withSpecifiedValues(String channelName,
-				Function<GrpcClientProperties, NamedChannel> channelFromProperties) {
+				Function<GrpcClientProperties, ChannelConfig> channelFromProperties) {
 			Map<String, String> map = new HashMap<>();
 			var propPrefix = "spring.grpc.client.%s.".formatted(channelName);
 			map.put("%s.address".formatted(propPrefix), "static://my-server:8888");
@@ -138,9 +149,9 @@ class GrpcClientPropertiesTests {
 		@Test
 		void withoutKeepAliveUnitsSpecified() {
 			Map<String, String> map = new HashMap<>();
-			map.put("spring.grpc.client.default-channel.idle-timeout", "1000");
-			map.put("spring.grpc.client.default-channel.keep-alive-time", "60000");
-			map.put("spring.grpc.client.default-channel.keep-alive-timeout", "5000");
+			map.put("spring.grpc.client.default-channel.idle-timeout", "1");
+			map.put("spring.grpc.client.default-channel.keep-alive-time", "60");
+			map.put("spring.grpc.client.default-channel.keep-alive-timeout", "5");
 			GrpcClientProperties properties = bindProperties(map);
 			var defaultChannel = properties.getDefaultChannel();
 			assertThat(defaultChannel.getIdleTimeout()).isEqualTo(Duration.ofSeconds(1));
@@ -157,6 +168,106 @@ class GrpcClientPropertiesTests {
 			var defaultChannel = properties.getDefaultChannel();
 			assertThat(defaultChannel.getMaxInboundMessageSize()).isEqualTo(DataSize.ofBytes(1000));
 			assertThat(defaultChannel.getMaxInboundMetadataSize()).isEqualTo(DataSize.ofBytes(256));
+		}
+
+	}
+
+	@Nested
+	class GetChannelAPI {
+
+		@Test
+		void withDefaultNameReturnsDefaultChannel() {
+			var properties = new GrpcClientProperties();
+			var defaultChannel = properties.getChannel("default");
+			assertThat(properties).extracting("defaultChannel").isSameAs(defaultChannel);
+			assertThat(properties).extracting("channels", InstanceOfAssertFactories.MAP).isEmpty();
+		}
+
+		@Test
+		void withKnownNameReturnsKnownChannel() {
+			Map<String, String> map = new HashMap<>();
+			// we have to at least bind one property or bind() fails
+			map.put("spring.grpc.client.channels.c1.enable-keep-alive", "false");
+			GrpcClientProperties properties = bindProperties(map);
+			var channel = properties.getChannel("c1");
+			assertThat(properties).extracting("channels", InstanceOfAssertFactories.MAP)
+				.containsExactly(entry("c1", channel));
+		}
+
+		@Test
+		void withUnknownNameReturnsNewChannelWithCopiedDefaults() {
+			var defaultChannel = new ChannelConfig();
+			defaultChannel.setAddress("static://my-server:9999");
+			defaultChannel.setDefaultLoadBalancingPolicy("custom");
+			defaultChannel.getHealth().setEnabled(true);
+			defaultChannel.getHealth().setServiceName("custom-service");
+			defaultChannel.setEnableKeepAlive(true);
+			defaultChannel.setIdleTimeout(Duration.ofMinutes(1));
+			defaultChannel.setKeepAliveTime(Duration.ofMinutes(4));
+			defaultChannel.setKeepAliveTimeout(Duration.ofMinutes(6));
+			defaultChannel.setKeepAliveWithoutCalls(true);
+			defaultChannel.setMaxInboundMessageSize(DataSize.ofMegabytes(100));
+			defaultChannel.setMaxInboundMetadataSize(DataSize.ofMegabytes(200));
+			defaultChannel.setUserAgent("me");
+			defaultChannel.getSsl().setEnabled(true);
+			defaultChannel.getSsl().setBundle("custom-bundle");
+			var properties = newProperties(defaultChannel, Map.of());
+			var newChannel = properties.getChannel("new-channel");
+			assertThat(newChannel).usingRecursiveComparison().ignoringFields("address").isEqualTo(defaultChannel);
+			assertThat(newChannel).hasFieldOrPropertyWithValue("address", "static://new-channel");
+			assertThat(properties).extracting("channels", InstanceOfAssertFactories.MAP).isEmpty();
+		}
+
+	}
+
+	@Nested
+	class GetTargetAPI {
+
+		@Test
+		void channelWithStaticAddressReturnsStrippedAddress() {
+			var defaultChannel = new ChannelConfig();
+			var channel1 = new ChannelConfig();
+			channel1.setAddress("static://my-server:8888");
+			var properties = newProperties(defaultChannel, Map.of("c1", channel1));
+			assertThat(properties.getTarget("c1")).isEqualTo("my-server:8888");
+			assertThat(properties).extracting("channels", InstanceOfAssertFactories.MAP)
+				.containsExactly(entry("c1", channel1));
+		}
+
+		@Test
+		void channelWithTcpAddressReturnsStrippedAddress() {
+			var defaultChannel = new ChannelConfig();
+			var channel1 = new ChannelConfig();
+			channel1.setAddress("tcp://my-server:8888");
+			var properties = newProperties(defaultChannel, Map.of("c1", channel1));
+			assertThat(properties.getTarget("c1")).isEqualTo("my-server:8888");
+			assertThat(properties).extracting("channels", InstanceOfAssertFactories.MAP)
+				.containsExactly(entry("c1", channel1));
+		}
+
+		@Test
+		void channelWithAddressPropertyPlaceholdersPopulatesFromEnvironment() {
+			var defaultChannel = new ChannelConfig();
+			var channel1 = new ChannelConfig();
+			channel1.setAddress("my-server-${channelName}:8888");
+			var properties = newProperties(defaultChannel, Map.of("c1", channel1));
+			var env = new MockEnvironment();
+			env.setProperty("channelName", "foo");
+			properties.setEnvironment(env);
+			assertThat(properties.getTarget("c1")).isEqualTo("my-server-foo:8888");
+		}
+
+	}
+
+	@Nested
+	class CopyDefaultsAPI {
+
+		@Test
+		void copyFromDefaultChannel() {
+			var properties = new GrpcClientProperties();
+			var defaultChannel = properties.getDefaultChannel();
+			var newChannel = defaultChannel.copy();
+			assertThat(newChannel).usingRecursiveComparison().isEqualTo(defaultChannel);
 		}
 
 	}
